Moves relevant classes into subscription manager.
